+++ ChestsAnywhere/Framework/Containers/StorageFurnitureContainer.cs
@@ -29,7 +29,10 @@ internal class StorageFurnitureContainer : IContainer
     public ContainerData Data { get; }
 
     /// <inheritdoc />
-    public bool CanConfigureAutomate { get; } = false; // Automate doesn't support storage containers
+    public bool CanConfigureAutomateStore => false; // Automate doesn't support storage containers
+
+    /// <inheritdoc />
+    public bool CanConfigureAutomateTake => false; // Automate doesn't support storage containers
 
 
     /*********