+++ Automate/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Fixed hoppers not automated in 2.4.0+.
+
 ## 2.4.1
 Released 19 July 2025 for SMAPI 4.1.10 or later.