+++ ChestsAnywhere/docs/release-notes.md
@@ -2,7 +2,7 @@
 
 # Release notes
 ## Upcoming release
-* Updated chest options to support Automate 2.4.0+ configurable chests.
+* Updated chest options to support Automate 2.4.0+ configurable chest types.
 
 ## 1.28.2
 Released 13 July 2025 for SMAPI 4.1.10 or later.